Fix model
Fixes all classes in the model package.
