+++ src/main/java/nl/tudelft/contextproject/model/Drawable.java
@@ -8,25 +8,33 @@ import com.jme3.scene.Spatial;
  * Interface representing a Drawable object.
  * This object has a geometry that can be updated.
  */
-public interface Drawable {	
+public interface Drawable {
+
 	/**
 	 * Getter for the spatial of this Drawable.
-	 * @return The spatial representing this Drawable.
+	 *
+	 * @return
+	 * 		the spatial representing this Drawable
 	 */
-	public Spatial getSpatial();
+	Spatial getSpatial();
 	
 	
 	/**
 	 * Method used for testing with mocked spatial.
 	 * Sets the spatial of this instance to the specified spatial.
-	 * @param spatial The new spatial.
+	 *
+	 * @param spatial
+	 * 		the new spatial
 	 */
 	void setSpatial(Spatial spatial);
 	
 	/**
 	 * Draw the Drawable on a map.
-	 * @param g The Graphics2D to draw on.
-	 * @param resolution The resolution of a single tile on the map.
+	 *
+	 * @param g
+	 * 		the Graphics2D to draw on
+	 * @param resolution
+	 * 		the resolution of a single tile on the map
 	 */
-	public void mapDraw(Graphics2D g, int resolution);
+	void mapDraw(Graphics2D g, int resolution);
 }
