+++ src/main/java/nl/tudelft/contextproject/model/Inventory.java
@@ -3,13 +3,15 @@ package nl.tudelft.contextproject.model;
 import java.util.ArrayList;
 
 import com.jme3.math.ColorRGBA;
+import nl.tudelft.contextproject.model.entities.Bomb;
+import nl.tudelft.contextproject.model.entities.Entity;
+import nl.tudelft.contextproject.model.entities.Key;
 
 /**
  * Class representing the players inventory.
- *
  */
 public class Inventory {
-	ArrayList<Entity> array; 
+	ArrayList<Entity> array;
 	private int keys;
 	private int bombs;
 
@@ -24,8 +26,9 @@ public class Inventory {
 
 	/**
 	 * Adds a key to the inventory.
+	 *
 	 * @param key
-	 * 		The key to be added
+	 * 		the key to be added
 	 */
 	public void add(Key key) {
 		array.add(key);
@@ -34,8 +37,9 @@ public class Inventory {
 	
 	/**
 	 * Adds a bomb to the inventory.
+	 *
 	 * @param bomb
-	 * 		The bomb to be added
+	 * 		the bomb to be added
 	 */
 	public void add(Bomb bomb) {
 		array.add(bomb);
@@ -44,8 +48,9 @@ public class Inventory {
 	
 	/**
 	 * Removes a key or bomb from your inventory.
+	 *
 	 * @param ent
-	 * 		The bomb/key to remove
+	 * 		the bomb/key to remove
 	 */
 	public void remove(Entity ent) {
 		if (ent instanceof Bomb) {
@@ -57,6 +62,7 @@ public class Inventory {
 				}
 			}
 		}
+
 		if (ent instanceof Key) {
 			for (Entity entity : array) {
 				if (entity instanceof Key) {
@@ -69,27 +75,28 @@ public class Inventory {
 				}
 			}
 		}
-		return;
 	}
 	
 	/**
 	 * Returns a bomb from the inventory.
+	 *
 	 * @return
-	 * 		A bomb from the inventory
+	 * 		a bomb from the inventory if the inventory contains a bomb
 	 */
 	public Bomb getBomb() {
-		Bomb bomb = null;
 		for (Entity ent : array) {
 			if (ent instanceof Bomb) {
 				return (Bomb) ent;
 			}
 		}
-		return bomb;
+		return null;
 	}
 
 	/**
 	 * Checks if the inventory contains a bomb.
-	 * @return returns true if you have a bomb
+	 *
+	 * @return
+	 * 		true if the inventory contains a bomb
 	 */
 	public boolean containsBomb() {
 		return (bombs > 0);
@@ -97,7 +104,9 @@ public class Inventory {
 
 	/**
 	 * Checks if the inventory contains a key.
-	 * @return returns true if you have a key
+	 *
+	 * @return
+	 * 		true if the inventory contains a key
 	 */
 	public boolean containsKey() {
 		return (keys > 0);
@@ -105,15 +114,17 @@ public class Inventory {
 
 	/**
 	 * Checks if the inventory contains a certain color key.
+	 *
 	 * @param color
-	 * 		Color of the key
-	 * 
-	 * @return returns true if you have the certain color key
+	 * 		color of the key
+	 * @return
+	 * 		true if the inventory contains a key of the wanted color
 	 */
 	public boolean containsColorKey(ColorRGBA color) {
 		if (!containsKey()) {
 			return false;
 		}
+
 		for (Entity ent : array) {
 			if (ent instanceof Key) {
 				Key key = (Key) ent;
@@ -122,14 +133,17 @@ public class Inventory {
 				}
 			}
 		}
+
 		return false;
 	}
 	
 	/**
 	 * Returns a key of the given color in the players inventory.
+	 *
 	 * @param color
-	 * 		The color of the key
-	 * @return The key of that color
+	 * 		the color of the key
+	 * @return
+	 * 		the key of that color
 	 */
 	public Key getKey(ColorRGBA color) {
 		for (Entity ent : array) {
@@ -139,14 +153,15 @@ public class Inventory {
 					return key;
 				}
 			}
-	}
+		}
+
 		return null;
 	}
 	
 	/**
 	 * Gives the size of the current inventory.
 	 * @return
-	 * 		Current size
+	 * 		the current size of the inventory
 	 */
 	public int size() {
 		return array.size();
