+++ src/main/java/nl/tudelft/contextproject/model/entities/Entity.java
@@ -1,8 +1,11 @@
-package nl.tudelft.contextproject.model;
+package nl.tudelft.contextproject.model.entities;
 
 import com.jme3.math.Vector3f;
 
 import nl.tudelft.contextproject.Main;
+import nl.tudelft.contextproject.model.Drawable;
+import nl.tudelft.contextproject.model.PhysicsObject;
+import nl.tudelft.contextproject.model.TickListener;
 
 /**
  * Abstract parent class for all entities.
@@ -13,15 +16,17 @@ public abstract class Entity implements Drawable, TickListener {
 
 	/**
 	 * Get the current state of this entity.
-	 * @return The current state.
+	 *
+	 * @return
+	 * 		the current state
 	 */
 	public EntityState getState() {
 		return state;
 	}
 
 	/**
-	 * Set the state of this entity to a new state.
-	 * @param newState The new state of this entity.
+	 * @param
+	 * 		newState The new state of this entity
 	 */
 	public void setState(EntityState newState) {
 		this.state = newState;
@@ -30,16 +35,23 @@ public abstract class Entity implements Drawable, TickListener {
 	/**
 	 * Move the entity by the specified amounts.
 	 * If the entity is a {@link PhysicsObject}, the physics location should also be moved.
-	 * @param x The amount of movement in the x-axis.
-	 * @param y The amount of movement in the y-axis.
-	 * @param z The amount of movement in the z-axis.
+	 *
+	 * @param x
+	 * 		the amount of movement in the x-axis
+	 * @param y
+	 * 		the amount of movement in the y-axis
+	 * @param z
+	 * 		the amount of movement in the z-axis
 	 */
 	public abstract void move(float x, float y, float z);
 
 	/**
 	 * Checks if this object is closer to the player than the specified distance.
-	 * @param dist The maximum distance to the player to detect the collision.
-	 * @return True if the player is closert than dist.
+	 *
+	 * @param dist
+	 * 		the maximum distance to the player to detect the collision
+	 * @return
+	 * 		true if the player is closert than dist
 	 */
 	public boolean collidesWithPlayer(float dist) {
 		Vector3f playerLoc = Main.getInstance().getCurrentGame().getPlayer().getSpatial().getLocalTranslation();
@@ -49,8 +61,9 @@ public abstract class Entity implements Drawable, TickListener {
 
 	/**
 	 * Get the location of an entity.
+	 *
 	 * @return
-	 * 			a Vector3f representing the location
+	 * 		a Vector3f representing the location
 	 */
 	public Vector3f getLocation() {
 		return this.getSpatial().getLocalTranslation();
