+++ src/main/java/nl/tudelft/contextproject/model/entities/EntityState.java
@@ -1,0 +1,10 @@
+package nl.tudelft.contextproject.model.entities;
+
+/**
+ * Enum representing the state of an Entity.
+ */
+public enum EntityState {
+	DEAD,
+	ALIVE,
+	NEW
+}
