+++ src/main/java/nl/tudelft/contextproject/model/level/DrawableFilter.java
@@ -3,7 +3,7 @@ package nl.tudelft.contextproject.model.level;
 import java.util.HashSet;
 import java.util.Set;
 
-import nl.tudelft.contextproject.model.Entity;
+import nl.tudelft.contextproject.model.entities.Entity;
 
 /**
  * A class representing a filter for drawing a map.
@@ -15,7 +15,9 @@ public class DrawableFilter {
 	
 	/**
 	 * Simple constructor with an empty whitelist.
-	 * @param hideUnexplored Boolean to indicate if explored tiles should be draw.
+	 *
+	 * @param hideUnexplored
+	 * 		boolean to indicate if explored tiles should be draw
 	 */
 	public DrawableFilter(boolean hideUnexplored) {
 		this(new HashSet<String>(), hideUnexplored);
@@ -23,8 +25,11 @@ public class DrawableFilter {
 	
 	/**
 	 * Normal constructor with a specific whitelist.
-	 * @param entitySet The whitelist to filter on.
-	 * @param hideUnexplored Boolean to indicate if explored tiles should be draw.
+	 *
+	 * @param entitySet
+	 * 		the whitelist to filter on
+	 * @param hideUnexplored
+	 * 		boolean to indicate if explored tiles should be drawn
 	 */
 	public DrawableFilter(Set<String> entitySet, boolean hideUnexplored) {
 		this.entitySet = entitySet;
@@ -33,8 +38,11 @@ public class DrawableFilter {
 
 	/**
 	 * Add an entity to the whitelist.
-	 * @param e The entity to add.
-	 * @return true when te entity is added, false if it was already in the list.
+	 *
+	 * @param e
+	 * 		the entity to add
+	 * @return
+	 * 		true when te entity is added, false if it was already in the list
 	 */
 	public boolean addEntity(Entity e) {
 		return entitySet.add(e.getClass().getSimpleName());
@@ -42,8 +50,11 @@ public class DrawableFilter {
 	
 	/**
 	 * Check if an entity is on the whitelist.
-	 * @param e Entity to check for.
-	 * @return true when the entity is on the whitelist, false otherwise.
+	 *
+	 * @param e
+	 * 		entity to check for
+	 * @return
+	 * 		true when the entity is on the whitelist, false otherwise
 	 */
 	public boolean entity(Entity e) {
 		return entitySet.contains(e.getClass().getSimpleName());
@@ -51,7 +62,9 @@ public class DrawableFilter {
 
 	/**
 	 * Check if unexplored tiles should be shown.
-	 * @return true if it should be shown, else otherwise.
+	 *
+	 * @return
+	 * 		true if it should be shown, false otherwise
 	 */
 	public boolean hideUnexplored() {
 		return hideUnexplored;
