+++ src/test/java/nl/tudelft/contextproject/model/InventoryTest.java
@@ -15,6 +15,9 @@ import com.jme3.material.MatParam;
 import com.jme3.material.MaterialDef;
 import com.jme3.math.ColorRGBA;
 
+import nl.tudelft.contextproject.model.entities.Bomb;
+import nl.tudelft.contextproject.model.entities.Door;
+import nl.tudelft.contextproject.model.entities.Key;
 import org.junit.Before;
 import org.junit.Test;
