+++ src/test/java/nl/tudelft/contextproject/model/entities/PlayerTriggerTest.java
@@ -1,10 +1,11 @@
-package nl.tudelft.contextproject.model;
+package nl.tudelft.contextproject.model.entities;
 
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.anyFloat;
 
+import nl.tudelft.contextproject.model.TickListener;
 import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.BeforeClass;
